tech(api): migrate campaign participation stats repository
