+++ api/tests/integration/infrastructure/repositories/campaign-participation-repository_test.js
@@ -634,266 +634,4 @@ describe('Integration | Repository | Campaign Participation', function () {
       });
     });
   });
-
-  describe('#getAllParticipationsByCampaignId', function () {
-    let campaignId;
-
-    beforeEach(async function () {
-      campaignId = databaseBuilder.factory.buildCampaign().id;
-      await databaseBuilder.commit();
-    });
-
-    it('returns an empty object when no participations', async function () {
-      await databaseBuilder.commit();
-
-      const result = await campaignParticipationRepository.getAllParticipationsByCampaignId(campaignId);
-
-      expect(result).to.deep.equal([]);
-    });
-
-    it('returns the list of the campaign', async function () {
-      databaseBuilder.factory.buildCampaignParticipation({ masteryRate: 0 });
-      databaseBuilder.factory.buildCampaignParticipation({ campaignId, masteryRate: 0, validatedSkillsCount: 0 });
-      await databaseBuilder.commit();
-
-      const result = await campaignParticipationRepository.getAllParticipationsByCampaignId(campaignId);
-
-      expect(result).to.deep.equal([{ masteryRate: '0.00', validatedSkillsCount: 0 }]);
-    });
-
-    it('returns the list of only isImproved=false participations', async function () {
-      databaseBuilder.factory.buildCampaignParticipation({
-        campaignId,
-        masteryRate: 0,
-        validatedSkillsCount: 0,
-        isImproved: false,
-      });
-      databaseBuilder.factory.buildCampaignParticipation({
-        campaignId,
-        masteryRate: 0,
-        validatedSkillsCount: 0,
-        isImproved: true,
-      });
-      await databaseBuilder.commit();
-
-      const result = await campaignParticipationRepository.getAllParticipationsByCampaignId(campaignId);
-
-      expect(result).to.deep.equal([{ masteryRate: '0.00', validatedSkillsCount: 0 }]);
-    });
-
-    it('returns the list of not deleted participations', async function () {
-      databaseBuilder.factory.buildCampaignParticipation({
-        campaignId,
-        masteryRate: 0,
-        deletedAt: null,
-        validatedSkillsCount: 0,
-      });
-      databaseBuilder.factory.buildCampaignParticipation({
-        campaignId,
-        masteryRate: 0,
-        validatedSkillsCount: 0,
-        deletedAt: new Date('2019-03-13'),
-      });
-      await databaseBuilder.commit();
-
-      const result = await campaignParticipationRepository.getAllParticipationsByCampaignId(campaignId);
-
-      expect(result).to.deep.equal([{ masteryRate: '0.00', validatedSkillsCount: 0 }]);
-    });
-  });
-
-  describe('#countParticipationsByStatus', function () {
-    describe('For assessment campaign', function () {
-      let campaignId;
-      let campaignType;
-
-      beforeEach(async function () {
-        const campaign = databaseBuilder.factory.buildCampaign();
-        campaignId = campaign.id;
-        campaignType = campaign.type;
-        await databaseBuilder.commit();
-      });
-
-      it('returns a default object when no participations', async function () {
-        await databaseBuilder.commit();
-
-        const result = await campaignParticipationRepository.countParticipationsByStatus(campaignId, campaignType);
-
-        expect(result).to.deep.equal({ started: 0, completed: 0, shared: 0 });
-      });
-
-      it("should not count any participation regardless of it's status when participation is deleted ", async function () {
-        databaseBuilder.factory.buildCampaignParticipation({ campaignId, status: SHARED, deletedAt: '2022-03-17' });
-        databaseBuilder.factory.buildCampaignParticipation({ campaignId, status: TO_SHARE, deletedAt: '2022-03-17' });
-        databaseBuilder.factory.buildCampaignParticipation({ campaignId, status: STARTED, deletedAt: '2022-03-17' });
-        await databaseBuilder.commit();
-
-        const result = await campaignParticipationRepository.countParticipationsByStatus(campaignId, campaignType);
-
-        expect(result).to.deep.equal({ started: 0, completed: 0, shared: 0 });
-      });
-
-      describe('Count shared Participation', function () {
-        it('counts participations shared', async function () {
-          databaseBuilder.factory.buildCampaignParticipation({ campaignId, status: SHARED });
-          await databaseBuilder.commit();
-
-          const result = await campaignParticipationRepository.countParticipationsByStatus(campaignId, campaignType);
-
-          expect(result).to.deep.equal({ started: 0, completed: 0, shared: 1 });
-        });
-
-        it('counts the last participation shared by user', async function () {
-          databaseBuilder.factory.buildCampaignParticipation({ campaignId, isImproved: true });
-          databaseBuilder.factory.buildCampaignParticipation({ campaignId });
-          await databaseBuilder.commit();
-
-          const result = await campaignParticipationRepository.countParticipationsByStatus(campaignId, campaignType);
-
-          expect(result).to.deep.equal({ started: 0, completed: 0, shared: 1 });
-        });
-      });
-
-      describe('Count completed Participation', function () {
-        it('counts participations completed', async function () {
-          databaseBuilder.factory.buildCampaignParticipation({ status: TO_SHARE });
-          databaseBuilder.factory.buildCampaignParticipation({
-            campaignId,
-            status: TO_SHARE,
-          });
-
-          await databaseBuilder.commit();
-
-          const result = await campaignParticipationRepository.countParticipationsByStatus(campaignId, campaignType);
-
-          expect(result).to.deep.equal({ started: 0, completed: 1, shared: 0 });
-        });
-
-        it('counts the last participations completed by user', async function () {
-          databaseBuilder.factory.buildCampaignParticipation({
-            campaignId,
-            status: TO_SHARE,
-            isImproved: true,
-          });
-          databaseBuilder.factory.buildCampaignParticipation({
-            campaignId,
-            status: TO_SHARE,
-          });
-
-          await databaseBuilder.commit();
-
-          const result = await campaignParticipationRepository.countParticipationsByStatus(campaignId, campaignType);
-
-          expect(result).to.deep.equal({ started: 0, completed: 1, shared: 0 });
-        });
-      });
-
-      describe('Count started Participation', function () {
-        it('counts participations started', async function () {
-          databaseBuilder.factory.buildCampaignParticipation({ status: STARTED });
-          databaseBuilder.factory.buildCampaignParticipation({
-            campaignId,
-            status: STARTED,
-          });
-
-          await databaseBuilder.commit();
-
-          const result = await campaignParticipationRepository.countParticipationsByStatus(campaignId, campaignType);
-
-          expect(result).to.deep.equal({ started: 1, completed: 0, shared: 0 });
-        });
-
-        it('counts the last participation started by user', async function () {
-          databaseBuilder.factory.buildCampaignParticipation({
-            campaignId,
-            status: STARTED,
-            isImproved: true,
-          });
-          databaseBuilder.factory.buildCampaignParticipation({
-            campaignId,
-            status: STARTED,
-          });
-
-          await databaseBuilder.commit();
-
-          const result = await campaignParticipationRepository.countParticipationsByStatus(campaignId, campaignType);
-
-          expect(result).to.deep.equal({ started: 1, completed: 0, shared: 0 });
-        });
-      });
-    });
-
-    describe('For profile collection campaign', function () {
-      let campaignId;
-      let campaignType;
-
-      beforeEach(async function () {
-        const campaign = databaseBuilder.factory.buildCampaign({ type: CampaignTypes.PROFILES_COLLECTION });
-        campaignId = campaign.id;
-        campaignType = campaign.type;
-        await databaseBuilder.commit();
-      });
-
-      it('returns a default object when no participations', async function () {
-        await databaseBuilder.commit();
-
-        const result = await campaignParticipationRepository.countParticipationsByStatus(campaignId, campaignType);
-
-        expect(result).to.deep.equal({ completed: 0, shared: 0 });
-      });
-
-      it("should not count any participation regardless of it's status when participation is deleted ", async function () {
-        databaseBuilder.factory.buildCampaignParticipation({ campaignId, status: SHARED, deletedAt: '2022-03-17' });
-        databaseBuilder.factory.buildCampaignParticipation({ campaignId, status: TO_SHARE, deletedAt: '2022-03-17' });
-        await databaseBuilder.commit();
-
-        const result = await campaignParticipationRepository.countParticipationsByStatus(campaignId, campaignType);
-
-        expect(result).to.deep.equal({ completed: 0, shared: 0 });
-      });
-
-      describe('Count shared Participation', function () {
-        it('counts participations shared', async function () {
-          databaseBuilder.factory.buildCampaignParticipation({ campaignId, status: SHARED });
-          await databaseBuilder.commit();
-
-          const result = await campaignParticipationRepository.countParticipationsByStatus(campaignId, campaignType);
-
-          expect(result).to.deep.equal({ completed: 0, shared: 1 });
-        });
-
-        it('counts the last participation shared by user', async function () {
-          databaseBuilder.factory.buildCampaignParticipation({ campaignId, isImproved: true });
-          databaseBuilder.factory.buildCampaignParticipation({ campaignId });
-          await databaseBuilder.commit();
-
-          const result = await campaignParticipationRepository.countParticipationsByStatus(campaignId, campaignType);
-
-          expect(result).to.deep.equal({ completed: 0, shared: 1 });
-        });
-      });
-
-      describe('Count completed Participation', function () {
-        it('counts participations completed', async function () {
-          databaseBuilder.factory.buildCampaignParticipation({ status: TO_SHARE });
-          databaseBuilder.factory.buildCampaignParticipation({ campaignId, status: TO_SHARE });
-          await databaseBuilder.commit();
-
-          const result = await campaignParticipationRepository.countParticipationsByStatus(campaignId, campaignType);
-
-          expect(result).to.deep.equal({ completed: 1, shared: 0 });
-        });
-
-        it('counts the last participation completed by user', async function () {
-          databaseBuilder.factory.buildCampaignParticipation({ campaignId, status: TO_SHARE, isImproved: true });
-          databaseBuilder.factory.buildCampaignParticipation({ campaignId, status: TO_SHARE });
-          await databaseBuilder.commit();
-
-          const result = await campaignParticipationRepository.countParticipationsByStatus(campaignId, campaignType);
-
-          expect(result).to.deep.equal({ completed: 1, shared: 0 });
-        });
-      });
-    });
-  });
 });
